Added PDF creator. Not fully implemented

+++ BattleNetShop/BattleNetShop.ReportsModel/ProductInformation.cs
@@ -6,7 +6,7 @@
 
         public int Quantity { get; set; }
 
-        public int Price { get; set; }
+        public double Price { get; set; }
 
         public string Vendor { get; set; }
